inventory_backend/requirements.txt
Django==4.2.7 
djangorestframework==3.14.0 
psycopg2-binary==2.9.7 
django-cors-headers==4.3.1 
python-dotenv==1.0.0 
